+++ src/server/controllers/logout.controller.ts
