+++ src/server/services/retrieveUser.service.ts
@@ -1,0 +1,9 @@
+/* eslint-disable no-underscore-dangle */
+import { User } from '../models/user';
+
+const retrieveUser = async (email: string) => {
+  const user = await User.findOne({ internalAccount: { email } });
+  return user;
+};
+
+export default retrieveUser;
